Package pygenn-suite recipe  parsable by conda-recipe-manager
