--- recipes/pygenn-cuda/meta.yaml
@@ -1,143 +1,0 @@
-{% set version = "5.3.0" %}
-
-package:
-  # Package name depends on the backend variant:
-  #  - backend == "cpu"  -> pygenn-cpu
-  #  - backend == "cuda" -> pygenn-cuda
-  name: pygenn-{{ backend }}
-  version: {{ version }}
-
-source:
-  url: https://github.com/genn-team/genn/archive/refs/tags/5.3.0_RC1.tar.gz
-  sha256: f1a6807d15f99d91732cd2484727100e54c39b055bcbe2126e7792d07c776fe3
-
-build:
-  number: 0
-  # CPU:  skip on osx
-  # CUDA: build only on linux-64 and win
-  skip: true  # [(backend == "cpu" and osx) or (backend == "cuda" and not (linux64 or win))]
-  string: "py{{ PY_VER | replace('.', '') }}{% if backend == 'cuda' %}_cuda{{ cuda_version | replace('.', '') }}{% endif %}_{{ PKG_BUILDNUM }}"
-  script: |
-    {% if backend == "cuda" %}
-    {% if win %}
-    :: --- CUDA backend build (Windows, modular CUDA) ---
-    :: Point GeNN at Conda's modular CUDA.
-    set "CUDA_PATH=%BUILD_PREFIX%\Library"
-    :: Let MSBuild projects (for libgenn/backends) link against modular CUDA without the VS CUDA plugin
-    set "CUDA_LIBRARY_PATH=%BUILD_PREFIX%\Library\lib"
-    :: Make sure modular CUDA headers/libs and DLLs are discoverable during build
-    set "INCLUDE=%BUILD_PREFIX%\Library\include;%INCLUDE%"
-    set "LIB=%BUILD_PREFIX%\Library\lib;%LIB%"
-    set "PATH=%BUILD_PREFIX%\Library\bin;%PATH%"
-    "%PYTHON%" setup.py install --single-version-externally-managed --record=record.txt
-    {% else %}
-    # --- CUDA backend build (Linux, modular CUDA) ---
-    export CUDA_PATH="${BUILD_PREFIX}"
-    export PATH="${BUILD_PREFIX}/bin:${PATH}"
-    "$PYTHON" setup.py install --single-version-externally-managed --record=record.txt
-    {% endif %}
-    {% else %}
-    {% if win %}
-    :: --- CPU-only backend build (Windows) ---
-    :: Disable CUDA
-    set CUDA_PATH=
-    "%PYTHON%" setup.py install --single-version-externally-managed --record=record.txt
-    {% else %}
-    # --- CPU-only backend build (Linux) ---
-    unset CUDA_PATH
-    "$PYTHON" setup.py install --single-version-externally-managed --record=record.txt
-    {% endif %}
-    {% endif %}
-  missing_dso_whitelist:
-    # Common private GeNN DLLs (bundled with PyGeNN)
-    - $RPATH/genn_Release_DLL.dll
-    - $RPATH/libffi_Release_DLL.dll
-    - $RPATH/genn_single_threaded_cpu_backend_Release_DLL.dll
-    # CUDA backend extras (only present in CUDA variant)
-    - $RPATH/libcuda.so.1                # Provided by NVIDIA driver at runtime (Linux)
-    - $RPATH/nvcuda.dll                  # Provided by NVIDIA driver at runtime (Windows)
-    - $RPATH/genn_cuda_backend_Release_DLL.dll
-
-requirements:
-  build:
-    - {{ stdlib("c") }}  # [linux]
-    - {{ compiler('cxx') }}
-    - make  # [linux]
-    - python
-    - setuptools
-    - pybind11
-    - psutil
-    # CUDA toolchain (CUDA variant only)
-    - cuda-nvcc {{ cuda_nvcc }}              # [backend == "cuda"]
-    - cuda-cudart-dev {{ cuda_cudart_dev }}  # [backend == "cuda"]
-    - cuda-libraries-dev {{ cuda_libraries_dev }}  # [backend == "cuda"]
-    - cuda-driver-dev {{ cuda_version }}     # [backend == "cuda" and linux]
-    - cuda-cccl {{ cuda_version }}           # [backend == "cuda"]
-
-  host:
-    - python
-    - numpy
-    - setuptools
-    - psutil
-    - pybind11
-    - libffi
-    # CUDA host deps (CUDA variant only)
-    - cuda-cudart-dev {{ cuda_cudart_dev }}  # [backend == "cuda"]
-    - cuda-driver-dev {{ cuda_version }}     # [backend == "cuda" and linux]
-    - cuda-cccl {{ cuda_version }}           # [backend == "cuda"]
-
-  run:
-    - python
-    - numpy
-    - psutil
-    - libffi
-    - importlib-metadata  # [py<38]
-    # JIT toolchain (CPU & CUDA differ slightly)
-    - make  # [linux]
-    - cxx-compiler          # [backend == "cpu"]
-    - {{ compiler('cxx') }} # [backend == "cuda"]
-    # CUDA runtime stack (CUDA variant only)
-    - cuda-version =={{ cuda_version }}      # [backend == "cuda"]
-    - cuda-nvcc {{ cuda_nvcc }}              # [backend == "cuda"]
-    - cuda-cudart {{ cuda_cudart }}          # [backend == "cuda"]
-    - cuda-libraries {{ cuda_libraries }}    # [backend == "cuda"]
-    - cuda-libraries-dev {{ cuda_libraries_dev }}  # [backend == "cuda"]
-    - cuda-driver-dev {{ cuda_version }}     # [backend == "cuda" and linux]
-    - cuda-cccl {{ cuda_version }}           # [backend == "cuda"]
-
-test:
-  imports:
-    - pygenn
-  files:
-    - tests/run_pygenn_cpu_smoketest.py   # [backend == "cpu"]
-    - tests/run_pygenn_cuda_smoketest.py  # [backend == "cuda"]
-  commands:
-    - python tests/run_pygenn_cpu_smoketest.py   # [backend == "cpu"]
-    - python tests/run_pygenn_cuda_smoketest.py  # [backend == "cuda"]
-
-about:
-  home: https://github.com/genn-team/genn
-  license: LGPL-2.1-or-later
-  license_family: LGPL
-  license_file: LICENSE
-  summary: >
-    {% if backend == "cuda" -%}
-    GeNN Python bindings (pygenn) with CUDA backend support
-    {%- else -%}
-    CPU-only PyGeNN (single-threaded backend, no CUDA required)
-    {%- endif %}
-  description: |
-    GeNN (GPU enhanced Neuronal Networks) is a code generation framework for
-    accelerating spiking neural network simulations using GPUs and CPUs.
-    This package provides the Python bindings (PyGeNN) and JIT compilation
-    support. The specific backend included in this build is selected via the
-    `backend` build variant:
-      - backend == "cpu":  single-threaded CPU backend only.
-      - backend == "cuda": CUDA-enabled backend with GPU support.
-  dev_url: https://github.com/genn-team/genn
-
-extra:
-  recipe-maintainers:
-    - Agrim-P777
-    - tnowotny
-    - neworderofjamie
