+++ recipes/pygenn-suite/conda_build_config.yaml
@@ -1,11 +1,4 @@
 # ----------------------------
-# Backend variants
-# ----------------------------
-backend:
-  - cpu
-  - cuda
-
-# ----------------------------
 # Python ABI
 # ----------------------------
 python:
@@ -50,7 +43,6 @@ cxx_compiler_version:
 # ----------------------------
 # CUDA toolchain & libraries
 # (single variant: 12.9)
-# Used only when backend == "cuda"
 # ----------------------------
 cuda_version:
   - 12.9
