+++ recipes/pygenn-suite/tests/run_pygenn_cpu_smoketest.py
